[bookshelf] Add simple sql retry strategy.
Sql no_connection errors are retried for downloads in an effort to maintain forward progress.
This makes the retry interval and number of attempts configurable, and defaults to never retrying.

This adds two new configuration values:
bookshelf:sql_retry_delay (the number of ms to delay) and
bookshelf:sql_retry_count (the number of times to retry).

If sql_retry_count is 0, we do not retry.

A future iteration might want to move this lower in the stack.