move dvm fs sync exclude list to a file
Instead fo maintaining a long list in a more general configuration file,
we'll use the --exclude-file feature of rsync and provide a list that
way.

Note that we still support extending this via a custom exclude-list
in the local config.yml.  If it turns out that there isn't a case for
one-off exclusions that don't belong in the exclusion file, we can
remove that down the road