[bookshelf] Avoid connection-timeout related test failures
For the tests that exercise high-concurrency situation, we want to
maintain a high queue size to ensure that the tests don't fail because
of capacity issues during the test.

Also, previously the sql_retry_count wasn't being initialized with the
configured retry count and was thus undefined; which caused problems in
the maybe_retry pattern match.

Set the default to match the default setting for the
application, which is zero retries.