[bookshelf] Expose sql_retry config
Bookshelf attempts to retry some failures from no_connections. The
intent of this is to salvage work in flight for requests that make
many connections; so this is confined to the chunk streaming code.

Two parameters control the behavior:
sql_retry_count, which limits the number of retries, and
sql_retry_delay, which is the number of ms to wait between retries.
The default is to do 0 retries, which turns the feature off, and relies on the pooler queue to handle overload conditions.

A future version of pooler that allowed the caller to somehow be
prioritized based on the work at risk would eliminate the need for
this feature.