Make run_sqitch work services beyond erchef
Partybus.run_sqitch was previously hard-coded
to apply database migrations only to opscode_chef,
from the opscode_chef schema path. It also
re-applied the legacy 1.0.4 schema with every new schema change.

This change updates the function so that it only applies the migration
it's told to apply. It also expands it to permit other services
to invoke run_sqitch for updating their own databases.

The partybus migrations were retrofitted to use the new form.

In addition migration 014 was updated to explicitly apply the
legacy 1.0.4 schema.

This was added as a precaution -- there should be no path that
has migration 14 (or later) running, that does not already have
the baseline sqitch deployments applied:
 * new installations will already have the baseline properly applied.
 * EC 11.current will already have the baseline
   properly applied, as sqitch was introduced sometime early in the
   11.x timeframe.
 * Upgrading from 11.old (pre-sqitch) requires an upgrade to
   11.x current before proceeding.