expand release process doc post-release steps, includes pin of Chef Client
These changes to post-release process will ensure the various components
and dependencies bundled with Chef Server are kept up to date with the
latest security and fixes.  This will also ensure that these updates are
done before work on the new release begins, so that we work through a
full release cycle is done with the updates in place.

This also locks the bundled Chef Client to the currently released stable
version at the start of development for a new release.