Added tests for extending read_access_group to allow access to clients and users within an org they are a member of (see below).
We are adding the ability for clients to access all other clients within the same org(s) as itself to make Chef Vault more functional.

I refactored the GET /organizations/:org/clients/:client endpoint to reflect the soon-to-be-simplified state of that endpoint:
1) Any user or client in the org can access it.
2) Any user or client not in the org cannot access it (the only exception being pivotal).
Given this simplified state, there are only 4 tests we need for that endpoint.

I also updated the /organizations/:org_name/users/:name endpoint tests since we will be extending read access to clients for users within the same org.