Added /orgs/org/users/user/keys(/key) endpoints and changed default perms on org scoped key GETs.
Now, all org scoped key GETs will allow access to any requestor within the same org (and the superuser).
The endpoints affected are
+ /orgs/org/clients/client/keys
+ /orgs/org/clients/client/keys/key
+ /orgs/org/users/user/keys
+ /orgs/org/users/user/keys/key
of which the latter two are new endpoints.

The new permissions where implemented by overriding default permissions for those GETs and checking if
the requestor is a memeber of the public_key_read_access group instead. This is a new group that, by default,
contains the users and clients group. Org admins can edit who is in this group, which changes who is allowed
to view the org scoped user public keys.

The existing user key endpoints
+ /users/user/keys
+ /users/user/keys/key
were not modified and contain the same permissions as beforehand.

Changelog-Entry: Added /orgs/org/users/user/keys(/key) endpoint and changed default perms on org scoped key GETs.

The following endpoints' GET methods can now be accessed by any requestor that is a member of the same organization:
+ /organizations/:org/clients/:client/keys
+ /organizations/:org/clients/:client/keys/:key
+ /organizations/:org/users/:user/keys
+ /organizations/:org/users/:user/keys/:key

The above org-scoped user keys endpoints are new and access to them can be controlled by an admin by editing memebership
of the public_key_read_access group.