Added GET /groups/:group/transitive_member/actors/:actor endpoint for checking recursive membership.
This endpoint will tell you if an actor is a member of a passed group or if that actor is a member of a group is recursively a member of a passed group.

For example, if brian is a member of the group clients, and clients is a member of the group superusers,
/groups/superusers/transitive_member/actors/brian would return true.

It returns JSON like:
  {"is_member" => true or false}

The framework is laid for adding GET /groups/:group/transitive_member/groups/:group as well, but this code currently depends on a db function called groups_for_actor,
an a similar groups_for_group does not exist yet. Notes have been left in the code to this effect.

Changelog-Entry: [bifrost] Added GET /groups/:group/transitive_member/actors/:actor endpoint for checking recursively checking group membership for a given actor on a given group.