Bump omnibus-software to pick up latest server-jre
omnibus-software now defaults the server-jre version to 8u74
so I've updated the Gemfile.lock to pick up this new version.
However, this newer version of omnibus-software no longer
contains the cacerts version we were overriding with (2014.08.20)
so I had to update that as well and chose to use the default.