Fix bug where requestor membership of public_key_read_access was not being properly tested for keys access.
Access to the org scoped user and client keys read endpoints are controlled by membership on the public_key_read_access group (which defaults to the users and clients groups); however, a bug resulted in testing membership on public_key_read_access for the member being requested instead of the requestor.

This commit fixes that bug and adds regression tests.