[omnibus] bypass_bootstrap? should ensure both creds exist
In addition to its use in the validator bypass_bootstrap? is called in
recipes/default.rb to determine whether or not we should run the
bootstrap cookbook.  The call to bypass_bootstrap? in recipes/default.rb
happens /after/ the private-chef-secrets.json file is written to disk.
Thus, this check was leading us to bypass the bootstrap even in cases
when this was the first node running reconfigure against an external
postgresql.

Additionally, given the current complexity of the bootstrap procedure, I
don't think we should actively use the `use_chef_backend` option for
anything but the "HAProxy routing enabled" bits until we ship the code
related to that feature.

Overall, I'm not thrilled with this solution. I've started to do some
code review of the bootstrap situation generally and will have a follow
up PR with some documentation on what I've found.