chef_elasticsearch:reindex no longer uses delete-by-query
The delete-by-query API was removed from Elasticsearch in 2.0. Since we
support external vanilla installations we cannot depend on the API, or
the corresponding plugin to exist. This change implements a form of
delete-by-query by using Elasticsearch's Scroll API to fetch all the
documents that match the query, and then delete them individually.

Signed-off-by: Tom Duffield <tom@chef.io>