Allow the ACLs endpoint to accept 'clients' and 'users'
With this change ACLs endpoint will accept 'clients' and 'users'
in the request body when 'actors' is present and is an empty list.

This change provides provides improvements to error handling - in any
case where an invalid actor or group is specified, the name(s) of any
invalid entities are included in the response.

Signed-off-by: Marc Paradise <marc@chef.io>