Add a 'detail=granular' query option to GET ACL endpoint
When specified, this option will split out users and clients within
the body response, and will force 'actors' to be empty.  This body
response is compatible with re-PUTing the body back to the object ACE
endpoint as-is, which will simplify changes to tooling when updating it
to use recently-added new ACL endpoint functionality.

Signed-off-by: Marc Paradise <marc@chef.io>