Don't 500 if search attribute is invalid
In the case where a user, when providing a search, provides an attribute
constraint (-a flag) that is invalid, we currently throw a 500. An
invalid attribute constraint most commonly is specifying a nested hash
using dot-notation when one or more values in that constraint are, in
fact, not a hash. This change prevents the system from throwing a 500
and presents the user with a message indicating that the attribute does
not exist.

For example, `knife search node "*:*" -a uptime.seconds` is invalid
because the key `'uptime'` does exist but it is a string, not a
hash. Thus, ej attempts to fetch the key `'seconds'` from that string,
throwing an error within ej for an invalid function clause. With this
change, the search will return "ERROR - Attribute does not exist".

Signed-off-by: Tom Duffield <tom@chef.io>