Test creation of an org by server admins (and regular users)
Server admins is now able to create orgs; add tests to cover the basic
permissions. We don't yet validate the details of org creation here.

Note: We lack API level methods for editing global groups, so we have
to use the chef-server-ctl command. This breaks running pedant against
a remote API until we have that method, but given that server-admins
is maturing as a feature it's important to be able to test.