A quick and dirty script to dump indexable data
This ctl-command creates json files in the working directory for indexable
data plus org data.  It will remove all attributes from
serialized_object (for those that have it). For data bag items it keeps
the keys but removes any value.

The intent is to extract a data set that can help us track down
reindexing consistency failures related to the object names and ids.

Signed-off-by: Marc Paradise <marc@chef.io>