Do not fail complete reindex when one object fails
There are cases when we do a reindex where the object name in the
serialized JSON does not match the name we get when selecting `(id, name)` from
the database.

Currently when this lookup fails, the entire reindex crashes. This
change makes it so that if we fail to find the name, instead of failing
entirely, we output a list of objects that failed to reindex to the
user.

Signed-off-by: Nathan L Smith <smith@chef.io>