Add validation endpoint to Chef Server.
We sometimes have the need to be able to determine if a request signed
with a client key is valid.  This adds an endpoint
/organizations/ORGNAME/validate/PATH that accepts a forwarded post,
and checks the signing as if it were sent to the PATH resource (PATH
can be any resource string). If the request validates, it returns 200,
and some potentially useful data. If it doesn't it returns 401 like
any other request with a bad signature.

Also includes:

* Refactor pedant and add test.

* Add raml docs for  validate endpoint.

  These are a first attempt at a yaml spec. However it isn't exact, as
  it isn't easy to to specify a resource that is multiple component,
  such as the /validate/... endpoint uses.

* tag tests as authentication

  ...because this really is what they do, and tests tagged with that are
  already excluded from running against chef-zero.

Signed-off-by: Mark Anderson <mark@chef.io>
Signed-off-by: Stephan Renatus <srenatus@chef.io>