Return correct number of rows when searching with Elasticsearch
Recent changes to the way we use the Elasticsearch paging APIs to do
Chef searches had caused searches to always return all rows when given a
`rows` query parameter.

This change reverts the normal search behavior to what it was before,
but uses the scroll API when doing reindexing.

It also adds Pedant tests to test that the `rows` parameter returns the
correct number of rows and the `start` parameter works correctly.

It also updates the reindexing pedant test to ensure that all rows have
been deleted, not just the first page.

(Note that our internal CI testing infrastructure only runs the Pedant
test suite against a default installation with Solr as the search
provider. These new tests should pass on either provider given the
changes included in this commit, or fail against an Elasticsearch-backed
earlier version.)

Signed-off-by: Nathan L Smith <smith@chef.io>