[bookshelf] Don't send Cache-Control headers for 404 responses
The Chef Backend tests run Chef Server with nginx cookbook caching
enabled. This led to test failures because of cached responses from
previous test cases being returned by nginx.

In a few cases, we found that nginx was caching 404 responses because
bookshelf was always sending a Cache-Control header. Since we never
want to cache 404s, the Cache-Control header is now removed from the
response before it is sent.

For a case where the problem was caused by a cached 200 response
(which we normally want to be cached), I've changed the test so that
it is unlikely that the tests are referring to the file.

Signed-off-by: Steven Danna <steve@chef.io>