Make oc-chef-pedant wait for queues to be empty in reindex tests
This possibly is one of our flaky tests -- we put 1010 items into the
index, they'll be put onto a queue, and we query Solr for the items we
expect to have moved out of the queue into Solr.

This way, we try to ensure that the rabbit queue is empty before
proceeding.

Signed-off-by: Stephan Renatus <srenatus@chef.io>