Revert ruby to 2.2.6
The goal is to move to 2.3.3 soon; however, since moving to 2.3.3
we've seen an increased rate of Timeout errors during particular
tests. We currently believe that this may be because of poor
interactions between our forked version of rest-client, our net/http
IPv6 patches, and the newer Ruby version.  Rolling this back for now
gives us more time to sort that out while keeping the build working.

Signed-off-by: Steven Danna <steve@chef.io>