Make chef_index_batch_tests a less flaky
Before, these often failed on travis -- and we've established that could
have been the case because the flush() we call in the test has been
executed _before_ the item as added. This change makes add_item wait for
the queue size to change to something non-empty. (This needs adjustments
if we add test cases that care for more than one item added and waited
for.)

Signed-off-by: Stephan Renatus <srenatus@chef.io>