[erchef] Improve reindex reliability when using ElasticSearch
The fire-and-forget reindex process make sense for
RabbitMQ+opscode-expander but does not make sense for the direct-to-ES
indexing methods.  In Hosted and other large environments, a reindex
may quickly overwhelm the HTTP connection pool, leading to reindex
failures that go unreported to the user.

This change collects all errors so that they can be reported at the
end of the reindex run. It also introduces retries with randomized
sleep to help reduce the chance of overall failures.

For large organizations, this new strategy may reduce the throughput
of reindexing. Such users may need to tune the reindex_batch_size and
the search_batch_max_wait parameters. Further, to finish reindexing in
a reasonable time, users may still need to adjust the HTTP connection
pool size and the batch size.

Signed-off-by: Steven Danna <steve@chef.io>