Use canned response openssl mock in oc_chef_wm tests
Before, the chef_keygen_cache worker pool started as part of
ensure_all_started would generate a set of RSA keys.

These are not really needed for these tests, and we can thus do without.
This change introduces a mock shell script to be used in place of
openssl, returning a fixed 2048bit key.

Signed-off-by: Stephan Renatus <srenatus@chef.io>