Support nonlocal group names
The server_admins group introduces an issue with ACLS and groups; we
have a global group in the context of a organization scoped
object. Not only do we have the potential of conflicts with a org
scoped group with the same name, we also lack a mechanisim to look up
groups that aren't in the org we're looking at.

This PR adds new syntax '::' to indicate scoping. ::GROUPNAME without
a prefix indicates a global (across multiple orgs) entity, while
ORGNAME::GROUPNAME refers to a group in an another org.

This should unblock the work around server admins, including improved
org creation.

This will need to be implemented in chef_zero to maintain feature parity.

Signed-off-by: Mark Anderson <mark@chef.io>