[chef_index] Fix racy chef_wait_group test
The old test was racy because it expected the gen_server to stop
faster than it would take to get to the erlang:is_process_alive call.
This wasn't always true in CI. Now, we install a monitor and wait for
it. If the gen_server doesn't stop, the test will timeout.

Signed-off-by: Steven Danna <steve@chef.io>