Update pooler-dependent apps to be manually started (erchef)
Updates erchef so that any application that depends on pooler
will have its startup deferred until after we know that chef_secrets
has been started.

This prevents a a dependency conflict where pooler will load
and create its defined pools as soon as it's loaded -- and in the
case of sqerl, that means it will try to connect to the DB immediately.

Because non-included applications will always have their deps started
before the main app (oc_erchef), pooler was always getting started
before chef_secrets.

The other approach to take here was to modify pooler to declare its
own dependency on chef_secrets - which is invalid, as the dependency is
indirectly introduced by a module using the dpooler app.

Another alternative is to omit pooler configuration from the application
conifguration so that it can start without starting the sqerl pool.  During
application startup, we can then manually start the pool.  This was
the original approach, and it may still be a more maintainable
alternative.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>