travis: cache oc_erchef plt
Rebar3 caches the PLT for the base version of erlang in
$HOME/.cache/rebar3. However, most of our time is spent adding 445
apps to this base PLT.  By setting the plt_location explicitly in the
rebar.config, we make it easier to cache this in travis and hopefully
save us a lot of time during the travis run.

I've opted to keep this cache inside the src/oc_erchef directory so
that the large number of apps we add to the PLT won't influence the
dialyzer runs for other erlang apps on your local workstation.

Signed-off-by: Steven Danna <steve@chef.io>