Do not set secrets ownership until after the owning user exists
When we initialize Veil on a new installation of Chef Server,
the opscode user has not yet been created.  For the time being,
we'll skip setting the ownership on calling Veil, and explicitly
set it as a file resource in recipe execution.

Important Note: this introduces a non-obvious ordering change.
If credentials.save is called after the private_keys.rb recipe,
the ownership will get reverted by veil.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>