[oc-chef-pedant] Don't directly depend on veil
Because oc-chef-pedant has a number of downstream dependencies, the
direct dependency on veil meant that veil was going to start leaking
into a number of other repositories and packages. I don't think we
want that until the API stabilizes a bit more and until it is a more
robust libary.

Signed-off-by: Steven Danna <steve@chef.io>