rabbitmq: add runit control script for termination
Some upstream change brought in by the upgrade has inhibited
chef-server-ctl stop rabbitmq from stopping the service, leaving the
beam process re-parented.

This change tells rabbitmq-server it was running under systemd, such
that it will not fork in to the background.

To ensure we don't kill the beam process brutally, a script was added
mimicking what the rabbitmq-server script unsuccessfully tried to do.

Signed-off-by: Stephan Renatus <srenatus@chef.io>