Change user-supplied secret logic to allow for any secret
Now, if for example chef-server.rb contains

    redis_lb['password'] = 'foo'

This would

1. be ingested in the secret store
2. trigger a warning if the value in the secret store is different from
   the one in the config
3. trigger a warning that it's in plain in chef-server.rb and should
   rather be taken from the secret store (i.e., removed from the config)

and could be set interactively via

    chef-server-ctl set-secret redis_lb password

or from the environment via

    REDIS_LB_PASSWORD=foobar chef-server-ctl set-secret redis_lb password

Replaces old ctl-commands by calls to set-secrets; with a whitelist.

Without the whitelisting, it would be quite easy to mistype the name of
a secret, and thus store it under the wrong key. Now, only anticipated
group/name combinations are allowed.

Note that the non-generic secret ingestion commands, i.e. those with
extra warnings and confirmations,

- postgresql/db_superuser_password
- rabbitmq/actions_password

are not retired, since `set-secret` does not support extras.

Signed-off-by: Stephan Renatus <srenatus@chef.io>
Signed-off-by: Steven Danna <steve@chef.io>