Install knife-ec-backup from git
This was previously pinned down because the upgrade procedure from
"Open Source Chef Server 11" uses knife-ec-backup and we didn't want
it to drift at all.  With the introduction of veil we either (1) need
to upgrade knife-ec-backup to handle veil or (2) turn on compatibility
mode if you are doing an upgrade from osc-11.

This change is an attempt to test 1.

Signed-off-by: Steven Danna <steve@chef.io>