Check all of a hash instead of early exiting
When comparing elements of a hash, if an element is an array we exit
if it matches, rather than continuing on to check the other keys.

This is because we used 'return true' in the body of a all? loop,
apparently intending to move to the next iteration of the loop, but
instead exiting the whole function with an incorrect 'true'.