Use higher compression level for release builds
When we've introduced xz compression for rpm and debs, level 1 was
picked as it took roughly as long as gz (which we had before), while
still giving us smaller artifacts.

However, if we do a release, it would make sense to use a higher level.
In general, the release build should come mostly from cache anyways, so
its build time should be smaller, so the inflation of that time would be
less terrible.

Level 6 is the default of the CLI tools for xz (xzcat etc).

Signed-off-by: Stephan Renatus <srenatus@chef.io>