Covert from ibrowse to httpc in oc_chef_wm_containers_SUITE
We think that the unreliable update test could be the result of a bug
in our old version of ibrowse.  Rather than hunting down that bug now,
we have converted this to httpc and plan to watch it over time.  If
the test becomes reliable we should consider an ibrowse update.

Signed-off-by: Steven Danna <steve@chef.io>