[dvm] use veil-env-helper where necessary
It's tedius to always prefix your invocations of `dvm start SRV` with
veil-env-helper, and it's even hard to remember which secrets are needed
for which service.

Also, this reverts the removal of Mixlib::ShellOut for #run_command,
since we now don't need to pass file descriptors through dvm calls.
However, the fixed calls to #exec have been kept as-is (there's no
harm in these).

Signed-off-by: Stephan Renatus <srenatus@chef.io>