Fix chef-server-ctl test for opsworks
The AWS Opsworks Automate offering has Chef Server using Automate's
elasticsearch instead of the built in Solr. This is done via the
external database feature of Chef Server.

This is implemented by setting the following in chef_server.rb:
opscode_solr4['external_url'] = http://localhost:8080/elasticsearch'

And then having nginx proxy that to the elasticsearch backend.

However that breaks pedant because use the solr_root helper, which
would take the above external_url and render http://localhost:8080.

If we use the solr_url helper instead, the full URI is rendered, and
the tests work.

Signed-off-by: Mark Anderson <mark@chef.io>