add libiconv dependency
This dependency is no longe rpulled in implicitly, but it's
absence was causing s390 builds to fail because libpopt
appeared to depend on it.  It was previously getting pulled in
by libxml2 which was building first; with recent omnibus software
updates it was no longer doing so.

A better fix may be to conditionally include the dependency
for s390x builds.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>