Move wait-for-rabbit functionality into runit's check script
Runit allows for adding check scripts that are executed to see if the
service is up and running. That way, after `sv start rabbitmq` (or
`/opt/opscode/init/rabbitmq start`) has finished successfully, we know
the service is up.

This removes the need for the wait-for-rabbit script.

Initially, it became apparent in #1220 that there are still situations
where rabbitmq has been started, but is not ready yet.

Signed-off-by: Steven Danna <steve@chef.io>
Signed-off-by: Stephan Renatus <srenatus@chef.io>