Do not return 400 for valid /users/USERNAME/_acl/PERM calls
Previously, if PUT was made to update a user ACL, the request would
return 400 if the request body contained a user. This was because the
code expected an OrgId to be set. Now, we assume that requests in the
global context (no org id), can only refer to users.

An alternative approach would be to extend the implementation of the
new scoping syntax so that it works everywhere.

A number of disabled pedant tests have been re-enabled.

Signed-off-by: Steven Danna <steve@chef.io>