Fix Chef 12 oc_id regression
The upgrade of oc_id to Chef 12 in commit

6824d90681a87ee18fb64f826507998978742d40

introduced 2 issues:

(1) Incorrect arguments to ServrAPI.new
(2) A change in ssl verification behavior

This commit

- Fixes issue (1),
- Make the ssl_verify_mode configurable
- Updates omniauth-chef to a version that contains similar fixes.

* SSL Verify Mode

The ssl verify mode has been set to verify_none. This isn't a behavior
change from previous versions since this was the default in the old
version of Chef we were previously pinned to.

Because we want to talk to oc_id via 'localhost' to avoid cross-node
requests, changing this to verify_peer is problematic since, even if
we did the work to add /var/opt/opscode/nginx/ca (or the user-provided
cert location) to the trusted_cert_dir, the cert would likely still
mismatch since in the common case such certs ar for the `api_fqdn` and
not "localhost".

Signed-off-by: Steven Danna <steve@chef.io>