[oc-chef-pedant] Add compliance proxy tests
This adds tests for our nginx proxy config for compliance requests.
It starts a stub compliance server on the configured port. If the
requests are rewritten as expected, the stub service will return 200.

Since this requires the chef-server.rb to contain proxy config, we
disable these tests by default, but can enable them in our CI
configuration.

Signed-off-by: Steven Danna <steve@chef.io>