[pedant] Smaller timeout in wait_for_http_listener, style updates
These are minor updates based on code review feedback. Since the
HTTP::Get will block with a fairly long connect_timeout by default, it
is unlikely that we will hit the sleep in
wait_for_http_listener. However, if we do, we expect for startup to be
fast in most cases, so a smaller sleep makes sense.

Signed-off-by: Steven Danna <steve@chef.io>