Enable asset compilation at runtime.
Eenable compilation of missing assets, so that
requests to `all.svg` no longer 404.  This
dependency is defined and added in chef-web-core.
Building the assets in precompile generates everything includeing
all-FINGERPRINT.svg, however Rails is not using the fingerprint
in this case.

This is a simple fix after exploring the asset generation and
not finding a sane way to either a) get rails to use
the fingerprint for this resource ; or b) get it to load the
un-fingerprinted version out of the gem.

Since this is the only missing asset, no impact is expected by
enabling compilation of missing resources.

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>