[ctl-commands] Call reindex escript with absolute path
In 636b10dfa402ecbb55e598201b54ee80ae4f44aa the configuration for the
various erlang applications were changed to use `{include_erts,
false}` as a release build option. It appears that a side effect of
that change is that the start_clean.boot file in

/opt/opscode/embedded/service/opscode-erchef/bin

no longer points at the correct ERTS installation.  Since that boot
file is used automatically by erl when starting the command from that
directory, the reindex command was failing to launch.

This change avoid the problem by calling the reindex script with an
absolute path. Alternatives would include reverting the include_erts
change or determining how to configure rebar/relx such that the
produced start_clean.boot file is correct.

Signed-off-by: Steven Danna <steve@chef.io>