[erchef] Fix ACL updates when actor names includes '.'
Unlike users and groups, client names can contain the `.` character.
In fact, this character this likely since the client name is often
based on the node name which is based on the FQDN of the node.

In chef/chef-server#1159, we added the ability to add name-spaced
actors and groups to ACLs.  However, that changed validated the actor
names against the more restrictive user/group regular expression.

This change uses the standard NAME_REGEX for validation which includes
`.`.  It does mean that some invalid user and group names will pass
validation; however, presumably such users and groups will fail to be
found in the database since they can't be created.

Fixes #1274

Signed-off-by: Steven Danna <steve@chef.io>