[CON-638] Experimental support for external ES5
This adds elastic search 5 support to chef_index but this support is
experimental.  It does not take care of upgrading a current chef-server
from ES 2 -> ES 5.  It also does not ensure that all other chef-server
functionality such as the reporting or the chef-backend plugin correctly
work when ES 5 is used.  Using ES5 as the search engine is completely
unsupported in a customer environment.

For the dev environment, a new optional VM has been added to
dev/Vagrantfile.  To enable it, check out defaults.yml for the commented
out 'elasticsearch' section.  It supports bring ES2 or ES5 currently.

Signed-off-by: Kartik Null Cating-Subramanian <ksubramanian@chef.io>