Add optional ulimit for erchef
Add a chef-server.rb setting `opscode_erchef['memory_maxbytes']` which,
when set to a non-nil value, adds `-d $BYTES` to the `chpst` command
which sets the process environment for erchef. This setting limits the
data memory segment for the erchef process to the specified value, which
caps the amount of RAM erchef may use.

This setting is intended to be used for standalone topologies, such as
when Chef Automate and Chef Server are colocated on the same machine. In
these topologies, heavy search load can cause erchef to consume large
amounts of memory which then triggers memory related failures in other
components, such as Elasticsearch. The `memory_maxbytes` setting allows
the administrator to make erchef fail instead of the other components in
these overload situations, allowing the system to recover more rapidly.

This setting is inteded to act as an emergency "circuit breaker" for
excessive load situations and should be set at the highest value at
which it can be effective, i.e., the value at which the host system's
memory would be completely utilized. Future load testing efforts are
required to provide guidance around recommended values.

Signed-off-by: Daniel DeLeo <dan@chef.io>