[erchef] Compile regular expression with unicode options
This adds the following options to the regexes we compile for input
validation (from the erlang docs):

- unicode: The regular expression is specified as a Unicode charlist()
  and the resulting regular expression code is to be run against a
  valid Unicode charlist() subject.

- ucp: Specifies that Unicode character properties are to be used when
  resolving \B, \b, \D, \d, \S, \s, \W and \w. Without this flag, only
  ISO Latin-1 properties are used. Using Unicode properties hurts
  performance, but is semantically correct when working with Unicode
  characters beyond the ISO Latin-1 range.

Note that a Unicode charlist() includes, as far as I can tell, a UTF-8
encoded binary (type unicode_binary() in the unicode module docs).

Both options are mutually required for our existing tests with UTF-8
user-data to pass.

Signed-off-by: Steven Danna <steve@chef.io>