[pedant] Add sleep to re-indexing tests
My best guess for what is happening in Wilson is that we are hitting a
race between the queue being empty and expander posting the data to
solr.

I'm pretty skeptical of this given that the search that is failing is
often the /last/ search in a series of searches so there is already a
large amount of time between seeing an empty queue and doing to
failing search.

The other issue I'm investigating is whether our queues_empty?
function may be erroneously reporting empty queues, but perhaps this
will help in the meantime.

Signed-off-by: Steven Danna <steve@chef.io>