Add 9.2 data_dir removal to `chef-server-ctl cleanup`
I've moved that resource from post_11_upgrade_cleanup. While it would in
just the same way from _there_, this should make the intention clearer.

Also, we'll delete the old log files of 9.2 now.

The addition the the release notes would theoretically be unnecessary
since this is what the postinst[1] instructions say anyways. However, I
believe it doesn't hurt to point it out again.

[1] https://github.com/chef/chef-server/blob/master/omnibus/package-scripts/chef-server/postinst#L30-L52

Signed-off-by: Stephan Renatus <srenatus@chef.io>