Remove EPMD-related IPv6 preflight check
With the epmd patch added for OTP >= 18.3, the problem is gone: the epmd
we ship now _never implicitly adds an interfaces to listen on_.

Before, whenever *anything* was set via ERL_EPMD_ADDRESS, epmd would
listen on that address _and_ 127.0.0.1 and ::1. On system without ::1 on
the loopback interface, it would thus fail.

The setting we use is ERL_EPMD_ADDRESS=127.0.0.1 to have it listen ONLY
on that IP.

The change (extracted from a PR by @msantos) we've made to epmd is that
now, it only listens on the specified ERL_EPMD_ADDRESS.

Signed-off-by: Stephan Renatus <srenatus@chef.io>